requirements.txt
pypng==0.20220715.0
loguru==0.7.2
